Make integration tests organized in subfolders
