+++ packages/engine/tests/units/message/mod.rs
@@ -1,0 +1,3 @@
+use crate::run_test;
+
+run_test!(sending);
