+++ packages/engine/tests/units/state/edit/dot/mod.rs
@@ -1,0 +1,3 @@
+use crate::run_test;
+
+run_test!(number);
