+++ packages/engine/tests/units/state/edit/dot/number/src/globals.json
